Extract monitor into separate file
