+++ packages/miniplex-core/src/core.ts
@@ -1,6 +1,6 @@
 import { Bucket } from "@miniplex/bucket"
 import { id } from "@hmans/id"
-import { createQueue } from "@hmans/queue"
+import { Monitor } from "./monitor"
 export * from "@miniplex/bucket"
 
 export type Predicate<E, D extends E> =
@@ -335,55 +335,6 @@ export class Query<E> extends Bucket<E> implements IQueryableBucket<E> {
     } else if (!wanted && has) {
       this.remove(entity)
     }
-  }
-}
-
-export class Monitor<E> {
-  protected queue = createQueue()
-  protected queueDisconnect = createQueue()
-
-  constructor(
-    public readonly bucket: Bucket<E>,
-    protected readonly setup: ((entity: E) => void) | undefined = undefined,
-    protected readonly teardown: ((entity: E) => void) | undefined = undefined
-  ) {
-    this.connect()
-  }
-
-  connect() {
-    /* Setup all existing entities */
-    if (this.setup) {
-      for (const entity of this.bucket) {
-        this.queue(() => this.setup!(entity))
-      }
-
-      /* Setup new entities as they are added */
-      this.queueDisconnect(
-        this.bucket.onEntityAdded.subscribe((entity) => {
-          this.queue(() => this.setup!(entity))
-        })
-      )
-    }
-
-    /* Teardown entities as they are removed */
-    if (this.teardown) {
-      this.queueDisconnect(
-        this.bucket.onEntityRemoved.subscribe((entity) => {
-          this.queue(() => this.teardown!(entity))
-        })
-      )
-    }
-
-    return this
-  }
-
-  disconnect() {
-    this.queueDisconnect.flush()
-    return this
-  }
-
-  run() {
-    this.queue.flush()
   }
 }
