+++ packages/miniplex-core/src/index.ts
@@ -1,4 +1,4 @@
 export { Bucket } from "@miniplex/bucket"
-export { Query, World, Monitor } from "./core"
+export { Query, World } from "./core"
 export type { Strict, With, Without } from "./core"
 export { hasAnyComponents, hasComponents, hasNoComponents } from "./helpers"
